Rename extension packages
